Fix issue with errors not bubbling upwards for account creation
if any field error was found after the submitting the signup form,
no error was shown. When calling the VetEmptyFields funct
the value of slice of errors wasn't being returned. Made the mistake
of assuming a slice points to its head. This was easily fixed by
making sure to return the slice of errors.

changes made in accts
 - removed funct parameter rsltErr []error from VetEmptyFields
 - funct VetEmptyFields now returns []error

changes made in main
 - made minor tweaks to the call to VetEmptyFields given the changes above
 - added funct call to NewHTTPError, to allow for errors to be
   displayed after the redirect to the "/view" webpage