Fix the missing capital letter testcase
Fixed the "missing capital letter" testcase for the password
field. I removed the spaces in the error messages for the vetPassword
func.  Then, I changed the error message in Test_accountcreation to
match the one in the vetPassword func. Lastly, I changed the password
used to in the testcase to be lowercase.

Changes made in accts file
 - removed the spaces in the error messages for the vetPassword func.

Changes made in main_test

  Change made to "missing at least 1 capital letter" testcase

    - rewrote the error message to match the error message
      in vetPassword func in the accts file.

    - changed the password  to be all lowercase.

    - replaced "Have" with "Missing" in the name of the testcase.

Changes made in the main_test file