Rename replyid to parentid
Renamed the replyID to parentID for the posts struct.
Came to realize the replyID can be confused with the threadid
since they both can refer to a id for a collection of replies.
It was lacking context, that's why replyid was changed to
parentID.

Changes made to the Posts file
 - renamed the field replyID to parentID for the post struct.
 - changed references to reply_id to parent_id.