Fix for multiple posts pt1
Added the func UserPostsByID which returns a list of
all of the user's posts when given their userid.

In general, UserPostsByID func is part of a fix to allow for creating multiple
posts. Once its called by addPosts func in the main file,
the refresh template will get a rewrite to loop
through each user post. This is still a work in progress.

tldr: fix for multiple posts pt1
 - added func UserPostsByID to posts file which returns a list of  all the user posts
 - wip fix for addPosts handler in main file
 - wip rewrite of refreshposts template file to support looping multiple posts

changes to posts.go file
 - added func UUserPostsByID, returns ([]*Post, error)