Fix multiple posts pt2
"Refreshposts" and the "posts" template files confliceted with each other.
On user login, the "posts" template would load the latest post,
but lose the data after submitting 2 posts. Data was lost
between the addposts func and vetlogin func. The vetlogin func used
the posts template, but the addposts func used the refresh template.
Making sure the addposts and vetlogin funcs both used a single
template fixed the issue of the a single user post not
consistently showing.

To fix the issue of multiple posts not being shown required
the posts template file to take a map filled with posts and loop
over them.

Changes made to the Posts file
 - rewrote the posts file to loop over a map of user posts

Changes made to the main file

  - rewrote vetLogin func to render a map filled with the user's posts
    along with posts tempalte

  - rewrote addPosts func to render the posts template alongside a map
    filled with all the user's posts.