Add the username for each post pt2
Submitted posts now show the poster's username along with the post
body. On the backend, the username  was added to the posts table
in the database. Given the changes to posts table, the posts.go file
has been updated to include the username field in the post struct.
This also includes all the funcs that reference it in.
As for the template file, posts.tmpl, it now includes the username
field w/ layout constraints for better formatting.

Changes to drumstick.sql file
 - rewrote the the posts table to include the username column
  - username is a foreign-key to the username column in the user_account table
 - updated the constraints to include the username column

Changes to posts.go file

 - added username field of type string to the post struct

 - added func usernameByID that takes an integer id
   and returns the username of the poster.
   if the id doesn't exist an error is returned along
   with an empty string

 - updated the func CreatePosts to be able to
   insert the username field to the posts table.

Changes to posts.tmpl file
 - updated the posts range loop to include the username field
 - added separate divs to apply layout constraints for
   a post's username and content body for better formatting.