Alter posts table so its possible to delete posts
When deleting a post the plan was to use the thread_id, but the name
doesn't match what its was doing. The thread_id acts like post
id. Thats why the thread_id was replaced with id. To better match the
action to the column.

The parent_id column is temporarly dropped from the posts table.
The current focus is being able to delete posts. Not replies.

Changes made to drumstick.sql file
 - added the id column which represents the unique id for each post.
 - temporarily removed the parent_id and thread_id.
   replies and threading arent the focus yet.