Fix func userposts_by_id sql query
Currently a user can only create posts to itself, but the
sql query in the func userPostsByID selects all posts regardles of
user_id. This meant the user_id being passed to the func wasn't doing
anything. This has been fixed by changing the sql query so that it
returns posts based on the passed user_id.

Changes made to posts.go file
 func UserPostsByID
  - rewrote the sql query for pgxscan select to return posts for
    the specificed user_id