Fix the route handling for user login
While the loginform is a get request under the '/login' route,
the login info was being sent as a post request to '/posts' route.
This is wrong, the '/posts' route is only meant to be used after authenticating the user.
This has been fixed by having the login info be sent as post request
under the '/login' route. The login template file has been updated.
The func vetLogin now redirects to the posts file.

Changes made main file

 func vetlogin
   -  redirects to '/posts'
   -  its route handler is now a post request to '/login'

Changes made to login template file
 -  html form now to points to the route 'login'