Fix the route handling for posts
The route handling for posts was setup wrong.
The route "/posts" needed both a get and a post request.
The problem was the route to 'POST' '/posts' was previously being used
by the vetLogin func. After fixing the vetLogin func to use a post request to
'/login', the route to 'POST' '/posts' could be used. This opened up
the space to use both the get and post method for '/posts'.
 I was also able to remove the boiler plate in the func addPosts
 since it can redirect back to get request for '/posts'. As for the
 'POST' '/posts' I renamed the '/refresh' route for the addPosts func
 since it was a work around for lacking the 'POST' request to
 '/posts'.

Changes made to the main file

 - added the GET '/posts' to view the user's post pagee
   - added the func viewFeed to help with that

 - func addPosts
  - added the redirect to get route '/posts'
    - removed the boiler plate that’s already found in GET '/posts'

Changes made to posts template file
 - updated the form action to "/posts" to match the change
   to addPosts func.