Add the user's username to the posts page
Rewrote the user's post page description to be a welcome message that
includes their username. This was done to make it clear what user is
logged in when viewing posts from other users.

Changes made to the posts.go file
 - UsernameByID has been exported to be used in the main file

Changes made the main file
 func viewFeed
  - added a callback to the func UsernameByID to pass the user's username
    to be rendered in the posts template file

Changes made to posts.tmpl file
 - rewrote the the subtitle message to be a welcome message that
   includes the user's username