prevent the current user from deleting posts they don't own
In the previous version user's could delete each other's posts.
This isn't case anymore, proper checks against the user-id of the
loggedin user vs the user_id of the post are in place. If the user-ids
don't match an error is returned stating the user_id of the current
session holder doesn’t match that of post.

Changes to posts file
 - new func UserIDByPostID
    - takes a post-id as a paremeter and returns the user-id
      by querying the database

Changes to sessionmanager file

 - UserCustomClaims struct
    - added user-id int field to get the user-id field
      of the current logged in user

 - func CreateToken
   - added a userID parameter to pass the currently logged in user's
     UserID as a field in their custom claims.

Changes to main file

 - func accountCreation
   - updated the func call to sessionManager.CreateToken
     to match changes

  - func deletePosts
    - added check against the current-sesison-holder deleting
      posts they didn't create, returns an error if the user-id of
      the logged in user doesn't match the user-id of the post.