Add a new testutil pkg for testing database queries
The old way the program handled database queries was testing
them on the production database, which isn't the best practice,
a test database is needed. The point is to use transactions for unit
testing. Problem is, all the funcs that query the database need to
accept both transactions and pgx pool connections. It requires a rewrite
where these funcs need to take an interface that accepts both
connections for production and the test database.

New module testutils/testdb.go
 - will setup the conn. to the test database