Fix overuse of pools by using one shared pool in main
All the funcs that query the database each create a pool.
This is inefficient since a single pool can be shared between all of
them. This commit specifically creates a server struct that holds a
pool connection to be used as a method receiver for all route handlers
to pass the context between database query funcs and an echo object to
allow for grouping routes.

changes made to main
 - new struct server
   - fields: db -> *pgxpool.Pool, e -> echo.Echo
   - meant to passed a method receiver to route handlers
     so that a single shared pool and a context
     is passed to all funcs that query the database instead of the
     multiple pools needlessly being created currently.