Tweak the viewfeed route to accept a server receiver
The route handler viewFeed now takes a method receiver of type server
where it passes an existing pool and context created in the func main
to the UsernameByID func. This removes the need to have that create
the connection and context in The usernameByID func, so its been
rewritten. Now, it accepts a context, a type that implements querier
interface, pgxpool and an integer id. As explained in the previos commit
the querier interface is meant to allow for creating conenctions
using pgxpool for production and unit testing with transactions.

Changes in posts.go file
 - func UsernameByID

  - now takes as params a context.Context, an value that implements
    type Querier and an int id.

  - removed code to setup a new pool and context, since its now being
    shared from the main.go

Changes in main.go file

- func viewFeed
   - added server method receiver
   - updated func call to posts.UsernameByID to take the pgxpool
     connection and request context

- func main
   - viewFeed route now called using sever method receiver