Move the querier interface to the backend
The querier interface is meant to allow for database connections
using pools and transactions. It doesn't make sense for the interface
to exist inside only one of the many pkgs that query the database.
That’s the querier interface has been moved to the backend module.

Changes made to backend.go
 - moved the querier interface froms posts.go
   to backend.go

Changes made to posts.go
 - updated all refs to the querier interface