Rewrite the unit test func for make-posts-tree
The unit test for the func MakePostsTree was complicated.
By using the createPosts and ReplyToPost funcs an  integration test
was being created instead of using a simple slice *posts with the
expected replies to check against. The unit test passes or fails if  makePostsTree
returns a []*posts w/ the expected number of parent posts, number of
replies per parent, and content of the  reply for parent 2.