Improve the unit test func for reply-to-posts
While the sequences and tables are reset for each unit test,
the expected post-id was hard-coded. Its best to check the post-id of
the reply against the post-id of the parent directly for consistency.
Another issue that was found not checking if the reply can be read.
This was fixed by calling the UserPostByID func which in this case
returns the reply given is post-id. Then the content of the reply
stored in the database is checked against the reply content
of the func call to ReplyToPost.

Changes made to posts_test file
 - func test_replyToPosts
   - replaced the the hard-coded parentID values w/ Post.parentID
   - new if check btw the reply content stored in the database
     the func call to ReplyToPost content.