Rewrite the posts file's test input to be read as their own files
The test data which created 1 or 2 dummy accts have been split into 2
json files to be read, dummyAcct.json & dummyAccts.json respectively. The
reason for the change is hardcoded inputs makes for a painful debugging
experience. The other reason is to organize accts that are meant to
represent different test cases. This change will start by modifying
the unit-tests Test_UserPostsByUserID and Test_UserPosts.

Changes to posts folder
       new testdata/dummyAcct.json
       new testdata/dummyAccts.json

Changes to posts_test file
       func Test_UserPostsByUserID
         - rmv dummyAcct struct -> read dummyAcct.json
       func Test_UserPostsByUserID
         - rmv dummyAccts struct -> read dummyAccts.json