Split the testcases for a bad acct into multi json files
The testcases for the accountCreation func will no longer be embedded.
Instead, each testcase is its own json file inside the folder
'testdata/invalidAccts'. The reason for splitting the files to make it
easier to read and add test cases. For this initial commit,
testcases for fname and lname will be added. Also the embedded testcases
have been commented out and the order of the acct fields in the
anonymous struct have been rearranged for clarity too.

Changes made to the accts_test file
 func Test_accountcreation
  - rearranged the account field attributes for tests []struct
  - commented out embedded testcases

Changes made to accts folder
 split testcases for a bad acct into multiple json files
 - new file testdata/invalidAccts/case01_fname_lname_empty.json
 - new file testdata/invalidAccts/case02_fname_has_spaces.json
 - new file testdata/invalidAccts/case03_fname_has_nums.json
 - new file testdata/invalidAccts/case04_fname_has_symbols.json
 - new file testdata/invalidAccts/case05_lname_has_spaces.json
 - new file testdata/invalidAccts/case06_lname_hase_nums.json
 - new file testdata/invalidAccts/case07_lname_hase_symbols.json