fix an empty email returning multiple errs
When running the invalidAccts/case10 the empty email returned not
only an empty field error, but the rest of the invalid email errs too.
To fix this, an if check was added to the func vetEmailAddress in
accts.go. If an empty email is passed it returns an empty []error.

Changes to accts.go
  func vetEmailAddress
   - if check -> empty email - > []error{}