Rewrite the user-login test
This rewrite is to update the func Test_userlogin to use
a db transaction to reset the state of the test database
after the test finishes. It also makes sure the func creates a
account before attempting to login. The account data is read from
accts/testdata/dummy_acct.json.

Changes to accts/testdata folder
 - new dummy_acct.json -> account
   - meant to be read by the func Test_userlogin
    where its used to create a temp Acct for a valid user-cred testcase

Changes to accts_test file
  func Test_Userlogin
    - encapsulate the logic in database transaction
      to reset the database before the test

    -  for the valid testcase an account is read and created from
       from the file dummy_acct.json

    - testcases for invalid user-creds are not stored in json files
      b/c how small there

     - removed the use of http request since it wasn't being used