Fix infinite loop user-login test
Turns out reading the database while a write operation is taking
place under the same transaction is a recipe for a deadlock.
This problem was fixed by handling the transaction manually. Firs, the
tables have to be empty and their sequences reset. Then, the created
account is committed followed by running the table tests.
As for the if check to make sure proper errors were returned it
required the invalid usercredentials error to be set globally in
the accts file since the error objects wouldn't be the same.
Once this all the done the tables and sequences are reset again.

Changes to accts.go file
 - new InvalidUserCreds error -> "incorrect email or password."
   - updated all its refs

Changes to accts_test.go file
 - replaced local invalidUserCreds w/ Import InvalidUserCreds
 - rmv func resetAndTestTx
 - made sure at beginning and end of the func table rows are empty and
   sequences are reset
 -  manually commit created account first -> run subtests