Export the remaining errs for the accts file
Decided to make the remaining errors for the accts file
global so it can be imported by the accts_test file
which has a duplicate set of those same errs. It should make it
so that the use of error.Is results in true since I'm using the same
err object. Also the ErrReqSymbol error has been renamed,
ErrReqEmailSymbol for better clarity

Changes to accts file
  List of exported errs

   - general acct field errs
     - ErrEmptyField
     - ErrHasPunct
     - ErrHasNums
     - ErrHasSymbols

   - user-creds err
     - InvalidUserCreds

    - email errs
      - renamed ErrReqSymbol -> ErrReqEmailSymbol
      - ErrReqEndingAddr

    - pswd errs
      - EmptyPswd
      - PswdTooShort
      - PswdTooLong
      - PswdHasNoCapLetters
      - PswdHasNoDigits
      - PswdHasSymbols